+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Validate unsupported values passed to `Query::build()`
 - Stop adding default `Content-Length` headers to `multipart/form-data` parts to comply with RFC 7578 section 4.8
 
+### Removed
+
+- Dropped support for PHP 7.2 and 7.3
+
 ## 2.10.1 - 2026-05-20
 
 ### Fixed